DTO implementation

+++ src/main/java/org/minipiku/pandalhopperv2/Repository/MetroStationRepository.java
@@ -3,5 +3,5 @@ package org.minipiku.pandalhopperv2.Repository;
 import org.minipiku.pandalhopperv2.Entity.MetroStation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MetroStationRepository extends JpaRepository<MetroStation, Integer> {
+public interface MetroStationRepository extends JpaRepository<MetroStation, Long> {
 }
